비밀번호를 이용한 삭제 기능 추가

+++ mysql을 이용한 게시판 만들기 프로젝트/board/del.php
@@ -2,7 +2,19 @@
     include "lib.php";
 
     $idx = $_GET['idx'];
+    $pwd = $_GET['pwd'];
     $name = mysqli_real_escape_string($connect, $idx);
+    $pwd = mysqli_real_escape_string($connect, $pwd);
+
+    $query = "select * from sing_board where idx = '$idx' and pwd= '$pwd' "; //넘겨받은 idx값과 pwd값을 선택
+    $result = mysqli_query($connect, $query);
+    $data = mysqli_fetch_array($result);
+    if(!$data['idx']){
+        echo "<script> alert('비밀번호가 일치하지 않습니다. 삭제가 불가능합니다. '); history.back(1); </script> ";
+        //history.back(1);은 방금 전창으로 이동하는 메서드
+        exit;
+    }
+
 
     $query = "delete from sing_board where idx='$idx'";
